+++ little-games/main/src/main/scala/com/github/killjoyer/Main.scala
@@ -13,9 +13,9 @@ import com.github.killjoyer.modules.chat.SimpleChatModule
 import com.github.killjoyer.modules.echo.UsersHandler
 import com.github.killjoyer.modules.echo.UsersModule
 import com.github.killjoyer.repositories.impls.RuDbDictionaryRepository
+import com.github.killjoyer.services.application.impls.ChatsManagerLive
+import com.github.killjoyer.services.application.impls.UserEventsRouterLive
 import com.github.killjoyer.services.domain.impls.BullsAndCowsServiceLive
-import com.github.killjoyer.services.impls.ChatsManagerLive
-import com.github.killjoyer.services.impls.UserEventsRouterLive
 import tofu.logging.zlogs._
 import zio._
