+++ little-games/service-layer/src/main/scala/com/github/killjoyer/services/domain/traits/BullsAndCowsService.scala
@@ -1,6 +1,6 @@
-package com.github.killjoyer.services.traits
+package com.github.killjoyer.services.domain.traits
 
-import com.github.killjoyer.services.traits.BullsAndCowsService.BullsAndCowsResult
+import com.github.killjoyer.services.domain.traits.BullsAndCowsService.BullsAndCowsResult
 import io.circe.generic.JsonCodec
 import zio.Task
